getLongPathValueByIndex 도입하여, 상세페이지에 적절한 게시물 노출

+++ src/main/java/com/sbs/jsp/board/article/ArticleRepository.java
@@ -36,7 +36,7 @@ public class ArticleRepository {
     return id;
   }
 
-  public Article findById(int id) {
+  public Article findById(long id) {
     for(Article article : datum) {
       if(article.getId() == id) {
         return article;
